COMMON: Uplift SafeMutexedSeekableSubReadStream to common
